Add StatelessLLMWithTemplate class
Add a class for connecting to a stateless LLM over a standard
LLM API while using a template for formatting the prompt and
defining a stop token.

This is especially helpful when an LLM is not trained with an
<|im_end|> token, like ALPACA. When the <|im_end|> token does not
exist, the model may continue on to start speaking for the user
and have a conversation with itself.

I am using this with the llama_server from Llama.cpp. This should
work with other hosts, like oLlama, vLLM, Kobold.cpp, etc. by
setting the appropriate http endpoint. Please let me know if you
experience any issues with this.